humanoid_ppo/requirements.txt
torch>=2.1
tensorboardX
numpy
tqdm
pyyaml

